Handle new model convention (prefix + arbitrary property name).
The old convention was to encode/decode the obsel.props.foo property
as :hasFoo "value";
To preserve ascending compatibility, the approach is to keep this
method. If a property name has no : in it, then it follows the old
convention. If it has a : in it, then it follows the new one.